Ensure permissions for update/delete

+++ lib/litelist_web/controllers/for_sale_controller.ex
@@ -34,29 +34,56 @@ defmodule LitelistWeb.ForSaleController do
 
   def edit(conn, %{"id" => id}) do
     for_sale = Posts.get_for_sale!(id)
-    changeset = Posts.change_for_sale(for_sale)
-    render(conn, "edit.html", for_sale: for_sale, changeset: changeset)
+    if permission?(conn.assigns.current_neighbor, for_sale) do
+      changeset = Posts.change_for_sale(for_sale)
+      render(conn, "edit.html", for_sale: for_sale, changeset: changeset)
+    else
+      unauthorized_redirect(conn)
+    end
+
   end
 
   def update(conn, %{"id" => id, "for_sale" => for_sale_params}) do
     for_sale = Posts.get_for_sale!(id)
-
-    case Posts.update_for_sale(for_sale, for_sale_params) do
-      {:ok, for_sale} ->
-        conn
-        |> put_flash(:info, "For sale updated successfully.")
-        |> redirect(to: for_sale_path(conn, :show, for_sale))
-      {:error, %Ecto.Changeset{} = changeset} ->
-        render(conn, "edit.html", for_sale: for_sale, changeset: changeset)
+    if permission?(conn.assigns.current_neighbor, for_sale) do
+      case Posts.update_for_sale(for_sale, for_sale_params) do
+        {:ok, for_sale} ->
+          conn
+          |> put_flash(:info, "For sale updated successfully.")
+          |> redirect(to: for_sale_path(conn, :show, for_sale))
+        {:error, %Ecto.Changeset{} = changeset} ->
+          render(conn, "edit.html", for_sale: for_sale, changeset: changeset)
+      end
+    else
+      unauthorized_redirect(conn)
     end
   end
 
   def delete(conn, %{"id" => id}) do
     for_sale = Posts.get_for_sale!(id)
-    {:ok, _for_sale} = Posts.delete_for_sale(for_sale)
+    if permission?(conn.assigns.current_neighbor, for_sale) do
+      {:ok, _for_sale} = Posts.delete_for_sale(for_sale)
 
+      conn
+      |> put_flash(:info, "For sale deleted successfully.")
+      |> redirect(to: for_sale_path(conn, :index))
+    else
+      unauthorized_redirect(conn)
+    end
+
+  end
+
+  defp permission?(neighbor, resource) do
+    if neighbor.id == resource.neighbor_id do
+      true
+    else
+      false
+    end
+  end
+  
+  defp unauthorized_redirect(conn) do
     conn
-    |> put_flash(:info, "For sale deleted successfully.")
+    |> put_flash(:error, "Unauthorized.")
     |> redirect(to: for_sale_path(conn, :index))
   end
 end
